requirements.txt
aenum==3.1.15
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==3.7.1
appdirs==1.4.4
APScheduler==3.9.1
arabic-reshaper==2.1.4
argon2-cffi==20.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asn1crypto==1.5.1
astroid==2.4.2
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==3.3.0
blinker==1.6.2
boto3==1.24.82
botocore==1.27.82
branca==0.6.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.3
comm==0.1.3
contourpy==1.1.0
crispy-bootstrap4==2022.1
cryptography==40.0.1
cssselect2==0.7.0
cycler==0.11.0
dataclasses-json==0.5.9
debugpy==1.6.7
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.1
Django==3.2.15
django-auto-logout==0.5.1
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap4==23.1
django-brutebuster2==0.2.0
django-crispy-forms==2.0
django-ipware==6.0.0
django-pdfkit==0.3.1
django-preventconcurrentlogins==0.8.2
django-ratelimit==4.1.0
django-rest-knox==4.2.0
djangorestframework==3.12.4
dlib==19.24.2
dynaconf==3.1.11
entrypoints==0.3
exceptiongroup==1.0.0rc9
executing==1.2.0
exif==1.6.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastjsonschema==2.17.1
filelock==3.0.12
Flask==2.3.2
folium==0.14.0
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.3.3
future==0.18.2
gunicorn==21.2.0
h11==0.14.0
h5py==3.9.0
html5lib==1.1
hvac==1.0.2
idna==3.4
importlib-metadata==1.7.0
imutils==0.5.4
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isoduration==20.11.0
isort==5.4.2
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.4
jsonschema==3.2.0
jsonschema-specifications==2023.6.1
jupyter==1.0.0
jupyter-console==6.4.0
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==4.12.0
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.4.4
knox==0.1.14
lazy-object-proxy==1.4.3
loguru==0.6.0
lxml==4.9.1
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.6.1
mistune==3.0.1
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.5.3
nbconvert==7.7.3
nbformat==5.9.1
nest-asyncio==1.5.1
notebook==6.3.0
notebook_shim==0.2.3
numpy==1.24.2
O365==2.0.31
oauthlib==3.2.2
opencv-python==4.7.0.72
oscrypto==1.3.0
outcome==1.2.0
overrides==7.3.1
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
pdfkit==0.6.0
pickleshare==0.7.5
piexif==1.1.3
Pillow==9.2.0
platformdirs==3.8.1
plum-py==0.8.7
prometheus-client==0.9.0
prompt-toolkit==3.0.18
psutil==5.9.5
psycopg2==2.8.6
psycopg2-binary==2.9.6
pure-eval==0.2.2
pycparser==2.20
pydantic==1.10.7
pygame==2.5.2
Pygments==2.8.1
pyHanko==0.14.0
pyhanko-certvalidator==0.19.5
pyhcl==0.4.4
PyJWT==2.5.0
pylint==2.6.0
pyparsing==2.4.7
pypdf==3.7.1
PyPDF3==1.0.6
pypng==0.20220715.0
pyproj==3.6.0
pyrsistent==0.17.3
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-ipware==2.0.0
python-json-logger==2.0.7
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pywin32==302
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
qrcode==6.1
qtconsole==5.0.3
QtPy==1.9.0
ratelimit==2.2.1
referencing==0.29.1
reportlab==3.6.11
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.8.10
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.4.3
Send2Trash==1.8.2
sentinelhub==3.9.1
shapely==2.0.1
shortuuid==1.0.9
six==1.15.0
sklearn==0.0.post5
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.1
stack-data==0.6.2
stringcase==1.2.0
svglib==1.4.1
sweetify==2.3.1
terminado==0.9.3
testpath==0.4.4
threadpoolctl==3.1.0
tifffile==2023.7.4
tinycss2==1.1.1
toml==0.10.1
tomli==2.0.1
tomli_w==1.0.0
tornado==6.3.2
tqdm==4.64.1
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
twilio==7.14.1
typing-inspect==0.9.0
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
uri-template==1.3.0
uritools==4.0.0
urllib3==1.26.12
utils==1.0.1
utm==0.7.0
validators==0.20.0
virtualenv==20.0.30
virtualenvwrapper-win==1.2.7
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
whitenoise==6.6.0
widgetsnbextension==3.5.1
win32-setctime==1.1.0
wrapt==1.12.1
wsproto==1.2.0
xhtml2pdf==0.2.8
yarl==1.8.2
zipp==3.8.1
